+++ src/Molecules/Select/Trigger/Trigger.types.ts
