+++ src/index.ts
@@ -84,7 +84,7 @@ import ProgressIndicator from './Molecules/ProgressIndicator';
 import PromotionBanner from './Molecules/PromotionBanner';
 import Rating from './Molecules/Rating';
 import SegmentedControl from './Molecules/SegmentedControl';
-import { Select as NewSelect } from './Molecules/NewSelect';
+import { Select } from './Molecules/Select';
 import SelectionCard from './Molecules/SelectionCard';
 import ShowMoreButton from './Molecules/ShowMoreButton';
 import Slider from './Molecules/Slider';
@@ -140,7 +140,7 @@ export type {
 
 /** Runtime code */
 export {
-  NewSelect,
+  Select,
   ActionModal,
   Accordion,
   AccordionItem,
